requirements.txt
asgiref
Django
pillow
sqlparse
tzdata
reportlab
django-allauth
python-dotenvrequests
PyJWT
cryptography
